feat(cli-presentation) splitted wallets from keys into seperate commands, added version command
