+++ packages/tools/kadena-cli/src/keys/index.ts
@@ -1,11 +1,6 @@
-import { createChangeWalletPasswordCommand } from './commands/keysChangeWalletPassword.js';
 import { createDecryptCommand } from './commands/keysDecrypt.js';
-import { createDeleteKeysCommand } from './commands/keysDeleteWallet.js';
-import { createGenerateHdKeysCommand } from './commands/keysHdGenerate.js';
-import { createImportWalletCommand } from './commands/keysImportWallet.js';
 import { createListKeysCommand } from './commands/keysList.js';
 import { createGeneratePlainKeysCommand } from './commands/keysPlainGenerate.js';
-import { createGenerateWalletCommand } from './commands/keysWalletGenerate.js';
 
 import type { Command } from 'commander';
 
@@ -15,12 +10,7 @@ export function keysCommandFactory(program: Command, version: string): void {
   const keysProgram = program
     .command(SUBCOMMAND_ROOT)
     .description(`Tool to generate and manage keys`);
-  createGenerateWalletCommand(keysProgram, version);
-  createImportWalletCommand(keysProgram, version);
-  createGenerateHdKeysCommand(keysProgram, version);
   createGeneratePlainKeysCommand(keysProgram, version);
-  createChangeWalletPasswordCommand(keysProgram, version);
-  createDeleteKeysCommand(keysProgram, version);
   createDecryptCommand(keysProgram, version);
   createListKeysCommand(keysProgram, version);
 }
