+++ packages/tools/kadena-cli/src/wallets/commands/walletsWalletGenerate.ts
@@ -8,18 +8,18 @@ import {
 } from '@kadena/hd-wallet/chainweaver';
 
 import { WALLET_DIR } from '../../constants/config.js';
+import {
+  displayGeneratedWallet,
+  displayStoredWallet,
+} from '../../keys/utils/keysDisplay.js';
+import { getWallet } from '../../keys/utils/keysHelpers.js';
+import * as storageService from '../../keys/utils/storage.js';
 import { services } from '../../services/index.js';
 import type { CommandResult } from '../../utils/command.util.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { createCommand } from '../../utils/createCommand.js';
 import { globalOptions } from '../../utils/globalOptions.js';
 import { log } from '../../utils/logger.js';
-import {
-  displayGeneratedWallet,
-  displayStoredWallet,
-} from '../utils/keysDisplay.js';
-import { getWallet } from '../utils/keysHelpers.js';
-import * as storageService from '../utils/storage.js';
 
 /**
  * Generates a new key for the wallet.
