+++ packages/tools/kadena-cli/src/wallets/tests/generateHdKeys.test.ts
@@ -1,6 +1,6 @@
 import { assert, describe, expect, it } from 'vitest';
-import { generateHdKeys } from '../commands/keysHdGenerate.js';
-import { importWallet } from '../commands/keysImportWallet.js';
+import { generateHdKeys } from '../commands/walletsHdGenerate.js';
+import { importWallet } from '../commands/walletsImportWallet.js';
 
 describe('generate hd keys command', () => {
   it('should generate hd keys correctly correctly', async () => {
